+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+
+## [0.3.0] - 2020-07-28
+
+### Added
+- Empty WebView with example.com
+- Floating button opening the phone app of the mobile device with a configured phone number
+
+
 ## [0.2.0] - 2020-07-26
 
 ### Added